pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "RAG Document Chatbot"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.101.0",
    "streamlit>=1.48.1",
    "PyPDF2>=2.0.0",
    "python-docx>=0.8.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.20.0",
]